fix(ffi): return a fixed-width uint32_t from unit_count, not pointer-width size_t
Fixes #24.

`aloud_session_unit_count` was declared `size_t` in the header and `usize` in
Rust, while the Kotlin/JNA binding declared it `Long`. JNA maps Kotlin `Long`
to a 64-bit native integer unconditionally, but `size_t` is pointer-width: 8
bytes on arm64-v8a and **4 on armeabi-v7a**, which scripts/build-android-core.sh
explicitly builds. On that ABI the callee returns 32 bits while JNA reads 64, so
the high half was whatever happened to be in the register — correct on 64-bit
devices, garbage on 32-bit ones.

This is exactly the failure class the architecture claims to prevent: it
compiles clean in all four languages and only misbehaves at runtime, on one ABI.

Fixed in the contract rather than per-binding, so no binding has to guess a
width again:

- `ffi.contract.md` gains an explicit rule — integers in the C surface are
  fixed-width, never pointer-width — and the signature table is updated.
- The header declares `uint32_t` (and includes <stdint.h> instead of <stddef.h>).
- Rust returns `u32`, saturating via `try_from` rather than wrapping; a document
  with more than u32::MAX sentences is unreachable in practice.
- Swift widens to its native `Int` at the edge.
- Kotlin declares `Int` (32-bit under JNA) and drops the now-redundant
  `.toInt()`.

+++ native/aloud-tts/android/src/main/java/com/aloud/tts/AloudCore.kt
@@ -18,7 +18,15 @@ private interface AloudCoreLib : Library {
     fun aloud_core_version(): Pointer
     fun aloud_session_new(text: String): Pointer?
     fun aloud_session_free(session: Pointer?)
-    fun aloud_session_unit_count(session: Pointer?): Long
+    /**
+     * The C ABI declares `uint32_t`, so this MUST be a 32-bit `Int` — JNA maps
+     * Kotlin `Long` to a 64-bit native integer unconditionally. It previously
+     * declared `Long` against a `size_t` return, which is 4 bytes on
+     * armeabi-v7a: JNA read 8, so the high half was whatever happened to be in
+     * the register. Correct on arm64, garbage on 32-bit. See the "fixed-width"
+     * rule in contracts/ffi.contract.md.
+     */
+    fun aloud_session_unit_count(session: Pointer?): Int
     fun aloud_session_dispatch(session: Pointer?, commandJson: String): Pointer
     fun aloud_string_free(ptr: Pointer?)
 }
@@ -42,7 +50,7 @@ class AloudCore(text: String) {
     private val session: Pointer =
         LIB.aloud_session_new(text) ?: throw CoreException("INVALID_UTF8", "core rejected document text")
 
-    val unitCount: Int get() = LIB.aloud_session_unit_count(session).toInt()
+    val unitCount: Int get() = LIB.aloud_session_unit_count(session)
 
     /** Send a command object and decode the response into a [Snapshot]. */
     fun dispatch(command: JSONObject): Snapshot {
